// backend/src/common/enums/permission.enum.ts
export enum PermissionName {
  // Users module
  USER_VIEW = 'USER_VIEW',
  USER_CREATE = 'USER_CREATE',
  USER_UPDATE = 'USER_UPDATE',
  USER_DELETE = 'USER_DELETE',

  // Roles module
  ROLE_VIEW = 'ROLE_VIEW',
  ROLE_CREATE = 'ROLE_CREATE',
  ROLE_UPDATE = 'ROLE_UPDATE',
  ROLE_DELETE = 'ROLE_DELETE',

  // Permissions module
  PERMISSION_VIEW = 'PERMISSION_VIEW',
  PERMISSION_ASSIGN = 'PERMISSION_ASSIGN',

  // Attendance module
  ATTENDANCE_VIEW = 'ATTENDANCE_VIEW',
  ATTENDANCE_CREATE = 'ATTENDANCE_CREATE',
  ATTENDANCE_UPDATE = 'ATTENDANCE_UPDATE',

  // Visits module
  VISIT_VIEW = 'VISIT_VIEW',
  VISIT_CREATE = 'VISIT_CREATE',
  VISIT_UPDATE = 'VISIT_UPDATE',
  VISIT_DELETE = 'VISIT_DELETE',

  // Dashboard module
  DASHBOARD_VIEW = 'DASHBOARD_VIEW',
}